2d pie

+++ minigraph/portability/sys_types.h
@@ -1,17 +1,19 @@
-//
-// Created by hsiaoko on 2022/3/18.
-//
 #pragma once
 
 #include <sys/types.h>
+
 #ifdef _WIN32
 #include <basetsd.h>  // @manual
 #endif
 
+// using gid_t = unsigned;
+// using vid_t = unsigned;
+// using vdata_t = unsigned;
+// using edata_t = unsigned;
 using gid_t = unsigned;
-using vid_t = unsigned;
-using vdata_t = unsigned;
-using edata_t = unsigned;
+using vid_t = size_t;
+using vdata_t = size_t;
+using edata_t = size_t;
 
 #define VID_MAX (((unsigned)(-1)) >> 1)
 #define GID_MAX (((unsigned)(-1)) >> 1)
